test(openai): move schema transformer tests to profiles
