+++ tests/models/test_openai.py
@@ -2829,85 +2829,6 @@ def test_strict_schema():
     )
 
 
-def test_strict_schema_keeps_supported_patterns():
-    class MyModel(BaseModel):
-        simple_pattern: Annotated[str, Field(pattern='^my-pattern$')]
-
-    schema_transformer = OpenAIJsonSchemaTransformer(MyModel.model_json_schema(), strict=None)
-
-    assert schema_transformer.walk() == snapshot(
-        {
-            'properties': {'simple_pattern': {'pattern': '^my-pattern$', 'type': 'string'}},
-            'required': ['simple_pattern'],
-            'type': 'object',
-            'additionalProperties': False,
-        }
-    )
-    assert schema_transformer.is_strict_compatible is True
-
-    escaped_schema_transformer = OpenAIJsonSchemaTransformer(
-        {
-            'properties': {'escaped_literal': {'pattern': '\\(?=USD', 'type': 'string'}},
-            'required': ['escaped_literal'],
-            'type': 'object',
-        },
-        strict=None,
-    )
-    assert escaped_schema_transformer.walk() == snapshot(
-        {
-            'properties': {'escaped_literal': {'pattern': '\\(?=USD', 'type': 'string'}},
-            'required': ['escaped_literal'],
-            'type': 'object',
-            'additionalProperties': False,
-        }
-    )
-    assert escaped_schema_transformer.is_strict_compatible is True
-
-
-def test_strict_schema_removes_unsupported_regex_lookarounds():
-    json_schema: dict[str, Any] = {
-        'properties': {
-            'before': {'pattern': '(?<=USD)\\d+', 'type': 'string'},
-            'after': {'pattern': '\\d+(?=USD)', 'type': 'string'},
-            'negative_before': {'pattern': '(?<!USD)\\d+', 'type': 'string'},
-            'negative_after': {'pattern': '\\d+(?!USD)', 'type': 'string'},
-        },
-        'required': ['before', 'after', 'negative_before', 'negative_after'],
-        'type': 'object',
-    }
-
-    schema_transformer = OpenAIJsonSchemaTransformer(json_schema, strict=None)
-
-    assert schema_transformer.walk() == snapshot(
-        {
-            'properties': {
-                'before': {'pattern': '(?<=USD)\\d+', 'type': 'string'},
-                'after': {'pattern': '\\d+(?=USD)', 'type': 'string'},
-                'negative_before': {'pattern': '(?<!USD)\\d+', 'type': 'string'},
-                'negative_after': {'pattern': '\\d+(?!USD)', 'type': 'string'},
-            },
-            'required': ['before', 'after', 'negative_before', 'negative_after'],
-            'type': 'object',
-            'additionalProperties': False,
-        }
-    )
-    assert schema_transformer.is_strict_compatible is False
-
-    assert OpenAIJsonSchemaTransformer(json_schema, strict=True).walk() == snapshot(
-        {
-            'properties': {
-                'before': {'type': 'string', 'description': 'pattern=(?<=USD)\\d+'},
-                'after': {'type': 'string', 'description': 'pattern=\\d+(?=USD)'},
-                'negative_before': {'type': 'string', 'description': 'pattern=(?<!USD)\\d+'},
-                'negative_after': {'type': 'string', 'description': 'pattern=\\d+(?!USD)'},
-            },
-            'required': ['before', 'after', 'negative_before', 'negative_after'],
-            'type': 'object',
-            'additionalProperties': False,
-        }
-    )
-
-
 def test_native_output_strict_mode(allow_model_requests: None):
     class CityLocation(BaseModel):
         city: str
